Feature 45 02

+++ modules/projects/application/handler/RemoveProjectMemberHandler.php
@@ -3,6 +3,7 @@
 namespace modules\projects\application\handler;
 
 use modules\projects\application\command\RemoveProjectMemberCommand;
+use modules\projects\application\port\IProjectAccess;
 use modules\projects\domain\repository\IProjectUserRepository;
 use modules\projects\domain\valueObject\ProjectId;
 use modules\projects\domain\valueObject\UserId;
@@ -11,10 +12,14 @@ use RuntimeException;
 class RemoveProjectMemberHandler
 {
     private IProjectUserRepository $projectUserRepository;
+    private IProjectAccess $projectAccess;
 
-    public function __construct(IProjectUserRepository $projectUserRepository)
-    {
-        $this->projectUserRepository = $projectUserRepository;
+    public function __construct(
+        IProjectUserRepository $projectUserRepository,
+        IProjectAccess $projectAccess
+    ) {
+        $this->projectUserRepository    = $projectUserRepository;
+        $this->projectAccess            = $projectAccess;
     }
 
     public function handle(RemoveProjectMemberCommand $command): void
@@ -22,13 +27,14 @@ class RemoveProjectMemberHandler
         $projectId = new ProjectId($command->projectId);
         $userId = new UserId($command->userId);
 
+        if (!$this->projectAccess->canRemoveUser($command->removedBy, $command->userId, $command->projectId)) {
+            throw new RuntimeException('You are not allowed to remove this user from the project');
+        }
+
         $projectUser = $this->projectUserRepository->find($projectId, $userId);
         if ($projectUser === null) {
             throw new RuntimeException("User is not a member of this project");
         }
-
-        // Проверка прав: админ проекта может удалять участников
-        // TODO: проверить, что $command->removedBy является админом
 
         $this->projectUserRepository->remove($projectUser);
     }
